jpa/chap01-persistence-context/src/main/java/com/ohgiraffers/section03/entity/EntityLifeCycle.java
package com.ohgiraffers.section03.entity;

import com.ohgiraffers.section01.entitymanager.EntityManagerGenerator;
import jakarta.persistence.EntityManager;

public class EntityLifeCycle {
    private EntityManager entityManager;

    public Menu findMenuByMenuCode(int menuCode){
        entityManager = EntityManagerGenerator.getInstance();
        return entityManager.find(Menu.class, menuCode);
    }

    // 확인 위한 반환 메소드
    public EntityManager getManagerInstance(){
        return entityManager;
    }
    //goto > test
}
